try to fix some pip install issues
OSX installs were failing if an openmp compatible compiler wasn't found,
backoff to using clang without openmp in this case
https://github.com/benfred/implicit/issues/1

On windows, we were passing in compile flags that weren't valid.
Visual Studo expects the '/W' flag to be an integer specifying the warning
level, and we were passing in a 'Wno-unused-function' to disable a warning.
https://github.com/benfred/implicit/issues/8 . Try to pass in the appropiate
options here