Fix issue with NaN appearing in CG code
https://github.com/benfred/implicit/issues/19#issuecomment-283164905

There was an issue with certain small matrices used as inputs causing NaN
values to appear in the CG optimizer. The reason seems to be a divide by zero
when rsold approached zero. Added a check to early exit in that case since the
optimization has succeeded at that point.