Add AnnoyAlternatingLeastSquares class
This class uses an Annoy Index to return the recommended items by
the inner product.

This involves transforming each vector by adding one extra dimension
as suggested in the paper:
"Speeding Up the Xbox Recommender System Using a Euclidean Transformation for
Inner-Product Spaces"
https://www.microsoft.com/en-us/research/wp-content/uploads/2016/02/XboxInnerProduct.pdf

Basically this involves transforming each feature vector so that they have the same
norm, which means the cosine of this transformed vector is proportional to the
dot product (if the other vector in the cosine has a 0 in the extra dimension).