Allow float32 inputs in ALS models
Previously all the ALS based models expected sparse matrices with a dtype of
numpy.float64 - and would fail to run in cython code if passed a float32
matrix instead. Fix this to allow passing either matrix type, and default
to float32 since the extra precision won't help in the majority of cases.