don't use c++11 for random number generation
Anaconda only supports VS 2008 (msvc9) for building python 2.7 on windows. This means
that we can't use C++11 features if we want to support this platform.

Rather than drop support for this, roll back this change to use the rand_r function
in the c stdlib.

Also enable support for 64 bit factors at the same time.