Make GPU ALS consistent with the CPU version
The CPU and GPU versions of the ALS model were returning slightly
different results, even when given the same random seed.

The difference was caused by order of operations: in a ALS
iteration the CPU version calculated user factors then
item factors and the GPU version did items then users.

Fix GPU code to be consistent with the the CPU version. After
this change the results are identical between the CPU/GPU
variants when given the same random state.