Better BPR Factor Initialization
Scale randomly initialized factors by the number of factors we're using. This leads
to substantial improvements in cross validation metrics like P@10. Also
center factors while we are at it. This approached was inspired by lightfm, and with
this changes the xval results are basically identical between these two libraries
for their BPR models.

With a 16 factor BPR model and learning_rate= 0.005 there is a
noticeable increase in p@10 with this change:

| Dataset        | P@10 Before Change | P@10 After Change |
| -------------- | ------------------ | ----------------- |
| Movielens 100k | 0.164              | 0.228             |
| Movielens 1M   | 0.186              | 0.243             |
| Movielens 10M  | 0.159              | 0.185             |
| Movielens 20M  | 0.142              | 0.228             |
| LastFM 360K    | 0.087              | 0.108             |