Fix GPU dot product when the # of factors wasn't warp aligned
Our MF algorithms produced incorrect results when the number of factors
wasn't warp aligned (like a multiple of 32).

This fixes a problem where we weren't correctly calculating the number of
warps in the dot function - meaning that we weren't including the results
of the last partial warp in the reduction.

Closes #262