Make approximate nearest neighbours models composable.
Approximate nearest neighbours used to only work for the ALS mode on the CPU.

This change makes it so that we can compose ANN methods with any matrix factorization
model (including BPR/LMF) and also use the GPU MF models as well.

Currently this provides the same api in implicit/approximate_als.py for backwards
compatability - but this may be removed at a future date.

Closes #487