Fix for recommend with 'items' filter
When using the 'items' filter on the recommend calls, we had two
different bugs in the latest code. One was that if you specified
a 'N' value greater than the size of the items you were ranking,
the results contained invalid values. The other was that we weren't
properly handling the 'filter_already_liked' option when the 'items'
option was set for MF models. Fix and add a unittest that would have
caught this