Use thrust::binary_search to verify negative samples on GPU
Use thrust::binary_search to verify negative samples on the GPU instead of
doing a linear scan in the BPR model.

This leads to a noticeable perf increase on larger datasets. For instance on the
Github stars dataset:
    * Using linear_search: 3.09s/it
    * using thrust::binary_search 1.53s/it
    * w/ verify_negative_samples=False  1.18s/it

This change doubles the BPR training performance on that dataset, and also
leads to times that are only 30% slower than not verifying samples at all.