use RAFT for select_k on the gpu
This changes to use RAFT https://github.com/rapidsai/raft for GPU
top-k code instead of faiss. The RAFT version is quite a bit faster,
and also doesn't have the same performance issues with small batch
sizes that faiss has (meaning we can delete a bunch of code that
was trying to work around that).