Add check_csr for partial_fit methods
We were seeing some poor result on partial_fit_*, that ended up
caused by passing a CSC instead of a CSR matrix.  (#682)

Add the same `check_csr` code to the partial_fit methods that is already
done in the fit method - this will warn if passed a non-csr matrix, and
automatically convert.