use threadpoolctl for detecting BLAS threadpools
We were using numpy.__config__.get_info to figure out what BLAS library was being used,
in an effort to warn people if they hadn't disabled the internal threadpool for the
BLAS they were using. This method is removed in numpy 1.26, causing errors.

Fix by using the threadpoolctl library to detect both the BLAS library and the number
of threads its configured for.

While we could automatically configure the BLAS library to reduce the threadpool size
to 1, this is process wide - and would have side effects for our users. Instead just
warn here, and give instructions for people on how to configure themselves.